refactor: 컨트롤러 경로 변경
