+++ app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller.legacy;
+package com.techcourse.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
