+++ app/src/main/java/com/techcourse/controller/UserSession.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller.legacy;
+package com.techcourse.controller;
 
 import java.util.Optional;
